backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "asyncpg>=0.30.0",
    "colorlog>=6.9.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "qrcode[pil]>=8.2",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.42",
    "sqlmodel>=0.0.24",
    "strawberry-graphql[fastapi]>=0.278.1",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]
